structure loops request done

+++ HarmonyDB.Index/HarmonyDB.Index.Analysis.Em/Models/Constants.cs
@@ -14,12 +14,12 @@ public static class Constants
     public static IReadOnlyList<(byte tonic, Scale scale)> Pairs { get; }
         = Indices.Select(x => ((byte)x.tonic, (Scale)x.scale)).ToList();
 
-    public static int GetMajorTonic((int tonic, Scale scale) scale)
+    public static int GetMajorTonic((byte tonic, Scale scale) scale)
     {
         return scale.scale == Scale.Major ? scale.tonic : GetParallelScale(scale).tonic;
     }
 
-    public static (byte tonic, Scale scale) GetParallelScale((int tonic, Scale scale) scale)
+    public static (byte tonic, Scale scale) GetParallelScale((byte tonic, Scale scale) scale)
     {
         return scale.scale == Scale.Major 
             ? ((byte)((scale.tonic - 3 + TonicCount) % TonicCount), Scale.Minor)
